refactor: move ext requests helpers to the router app
