+++ src/lib/api_app/src/api/routes/testbed/productizers/figure/mod.rs
@@ -1,9 +1,10 @@
 use std::env;
 
 use super::parse_testbed_request_headers;
-use crate::api::requests::post_json_request;
 
-use openapi_router::{requests::ParsedRequest, responses::APIResponse};
+use openapi_router::{
+    requests::post_json_request, responses::APIResponse, router::ParsedRequest,
+};
 
 pub mod figure_models;
 use figure_models::{PopulationQuery, PopulationResponse};
