+++ src/lib/api_app/src/api/routes/testbed/productizers/person/job_applicant_profile/mod.rs
@@ -1,5 +1,5 @@
 use openapi_router::{
-    requests::ParsedRequest,
+    router::ParsedRequest,
     responses::APIResponse,
 };
