+++ src/lib/api_app/src/api/routes/testbed/productizers/user/mod.rs
@@ -1,13 +1,10 @@
 use serde_json::{json, Value as JSONValue};
 
-use crate::api::{
-    requests::post_json_request,
-};
 use openapi_router::{
-    requests::ParsedRequest,
+    router::ParsedRequest,
+    requests::post_json_request,
     responses::APIResponse,
 };
-
 
 use super::{ parse_testbed_request_headers, build_data_product_uri };
