+++ src/lib/api_app/src/lib/openapi_router/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 openapi_router_derive = { path = "../openapi_router_derive" }
+utils = { path = "../utils" }
 
 serde_json = { workspace = true }
 serde = { workspace = true }
@@ -14,5 +15,4 @@ reqwest = { workspace = true }
 simple_logger = { workspace = true }
 lambda_http = { workspace = true }
 futures = { workspace = true }
-
-utils = { path = "../utils" }
+stopwatch = { workspace = true }
