+++ src/lib/api_app/src/lib/utils/src/environment.rs
@@ -1,0 +1,3 @@
+pub fn get_stage() -> String {
+    std::env::var("STAGE").unwrap_or_else(|_| "local".to_string())
+}
